requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.1.0
backports.zoneinfo==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
distro==1.9.0
Django==4.2.19
django-environ==0.11.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
lxml==5.3.1
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.64.0
orjson==3.10.15
packaging==24.2
pdf2image==1.17.0
pgvector==0.3.6
pillow==10.4.0
pip-autoremove==0.10.0
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-core==2.27.2
PyMuPDF==1.24.11
pytesseract==0.3.13
python-docx==1.1.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
tenacity==8.5.0
tqdm==4.67.1
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.3
yarl==1.15.2
